Move base test classes to valve_test_lib.py
